requirements.txt
beautifulsoup4==4.12.3
cliff==4.7.0
Cython==3.0.10
geopy==2.4.1
GitPython==3.1.43
GitPython==3.1.43
irrCAC==0.4.2
jieba==0.42.1
jsonlines==4.0.0
lxml==5.2.2
matplotlib==3.8.4
mediacloud==4.1.4
nltk==3.8.1
numpy==1.23.0
pandas==2.2.2
polyglot==16.7.4
pyecharts==2.0.6
python_igraph==0.11.6
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.4.2
scipy==1.14.0
seaborn==0.13.2
selenium==4.23.1
sentence_transformers==3.0.1
shap==0.46.0
spacy==3.7.5
spacy_langdetect==0.1.2
statsmodels==0.14.2
torch==2.3.1
tqdm==4.66.4
trafilatura==1.12.0
